sys: duplicate vaguely platform-specific code
